[API] Get joints in fn and joints constrained
Get pair of joints involved in differentiable fn

Get locked joint for explicit constraint

Add method to retrieve a joint whose whole or partial configuration will
decide the value of the implicit function in an explicit constraint.
This is currently done by inspecting the set of configuration input and output
intervals: if input has size 0 and output falls into interval of a joint,
then the joint is "locked" its config alone will affect the fn value.

Fix bug comparing joints from their pointers

Fix: locked joint should be locked wrt parent joint

Previous implementation misunderstands the locked joint, and assumes
that locked joint is locked wrt the world joint. The correct
understanding is that locked joints are locked wrt the parent joint,
which may or may not be the world joint.

Edit documentation based on comments

Clearer documentation on how the method works and what the arguments are

Add API check if all rows in constraint are active

Add API to find joint pair constrained

Return a pair of joints whose relative pose is fully constrained because
of constraint.

Clearer name for API to get joints involved

Add API for convex shape contact complement and hold
For now, it only works if the all convex shapes in floorConvexShapes_
are on the same joint an all the convex shapes in objectConvexShapes_
are on the same joint.

Refactor APIs and fix minor bugs